WIP: Fix bug causing line numbers to be swallowed
Fixes a bug causing line numbers to be swallowed for fatal errors in
ESLint. WIP since some tests are failing and they're a bit difficult to
understand, so I haven't been able to figure out what's going on in a
reasonable amount of time. Opening a WIP PR and issue to ask the author.